requirements.txt
--index-url https://nexus.pgk.ru/repository/pypi-group/simple

aiofiles==0.8.0 ; python_version >= "3.9" and python_version < "4.0"
alembic==1.10.2 ; python_version >= "3.9" and python_version < "4.0"
anyio==3.6.2 ; python_version >= "3.9" and python_version < "4.0"
asgi-ratelimit==0.10.0 ; python_version >= "3.9" and python_version < "4.0"
asgiref==3.6.0 ; python_version >= "3.9" and python_version < "4.0"
bcrypt==4.0.1 ; python_version >= "3.9" and python_version < "4.0"
bleach==6.0.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.9" and python_version < "4"
cffi==1.15.1 ; python_version >= "3.9" and python_version < "4.0"
chardet==5.1.0 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.1.0 ; python_version >= "3.9" and python_version < "4"
click==8.1.3 ; python_version >= "3.9" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0"
contourpy==1.0.7 ; python_version >= "3.9" and python_version < "4.0"
cryptography==3.4.8 ; python_version >= "3.9" and python_version < "4.0"
cx-oracle==8.3.0 ; python_version >= "3.9" and python_version < "4.0"
cycler==0.11.0 ; python_version >= "3.9" and python_version < "4.0"
docutils==0.19 ; python_version >= "3.9" and python_version < "4.0"
ecdsa==0.18.0 ; python_version >= "3.9" and python_version < "4.0"
et-xmlfile==1.1.0 ; python_version >= "3.9" and python_version < "4.0"
fastapi-viewsets==0.1.7 ; python_version >= "3.9" and python_version < "4.0"
fastapi==0.79.1 ; python_version >= "3.9" and python_version < "4.0"
fonttools==4.39.3 ; python_version >= "3.9" and python_version < "4.0"
fuzzywuzzy==0.18.0 ; python_version >= "3.9" and python_version < "4.0"
greenlet==2.0.2 ; python_version >= "3.9" and (platform_machine == "aarch64" or platform_machine == "ppc64le" or platform_machine == "x86_64" or platform_machine == "amd64" or platform_machine == "AMD64" or platform_machine == "win32" or platform_machine == "WIN32") and python_version < "4.0"
gunicorn==20.1.0 ; python_version >= "3.9" and python_version < "4.0"
h11==0.14.0 ; python_version >= "3.9" and python_version < "4.0"
idna==3.4 ; python_version >= "3.9" and python_version < "4"
importlib-metadata==6.1.0 ; python_version >= "3.9" and python_version < "4.0"
importlib-resources==5.12.0 ; python_version >= "3.9" and python_version < "3.10"
jaraco-classes==3.2.3 ; python_version >= "3.9" and python_version < "4.0"
jeepney==0.8.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "linux"
jinja2==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
keyring==23.13.1 ; python_version >= "3.9" and python_version < "4.0"
kiwisolver==1.4.4 ; python_version >= "3.9" and python_version < "4.0"
levenshtein==0.20.9 ; python_version >= "3.9" and python_version < "4.0"
mako==1.2.4 ; python_version >= "3.9" and python_version < "4.0"
markupsafe==2.1.2 ; python_version >= "3.9" and python_version < "4.0"
matplotlib==3.7.1 ; python_version >= "3.9" and python_version < "4.0"
more-itertools==9.1.0 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.24.2 ; python_version < "4.0" and python_version >= "3.9"
openpyxl==3.1.2 ; python_version >= "3.9" and python_version < "4.0"
orjson==3.8.9 ; python_version >= "3.9" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.9" and python_version < "4.0"
pandas==1.5.3 ; python_version >= "3.9" and python_version < "4.0"
passlib==1.7.4 ; python_version >= "3.9" and python_version < "4.0"
pillow==9.5.0 ; python_version >= "3.9" and python_version < "4.0"
pkginfo==1.9.6 ; python_version >= "3.9" and python_version < "4.0"
psycopg2-binary==2.9.5 ; python_version >= "3.9" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.9" and python_version < "4.0"
pycparser==2.21 ; python_version >= "3.9" and python_version < "4.0"
pydantic==1.10.7 ; python_version >= "3.9" and python_version < "4.0"
pydantic[dotenv]==1.10.7 ; python_version >= "3.9" and python_version < "4.0"
pygments==2.14.0 ; python_version >= "3.9" and python_version < "4.0"
pyparsing==3.0.9 ; python_version >= "3.9" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.0.0 ; python_version >= "3.9" and python_version < "4.0"
python-jose[cryptography]==3.3.0 ; python_version >= "3.9" and python_version < "4.0"
python-json-logger==2.0.7 ; python_version >= "3.9" and python_version < "4.0"
python-levenshtein==0.20.9 ; python_version >= "3.9" and python_version < "4.0"
python-multipart==0.0.5 ; python_version >= "3.9" and python_version < "4.0"
pytz==2023.3 ; python_version >= "3.9" and python_version < "4.0"
pywin32-ctypes==0.2.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "win32"
pyyaml==6.0 ; python_version >= "3.9" and python_version < "4.0"
rapidfuzz==2.15.0 ; python_version >= "3.9" and python_version < "4.0"
readme-renderer==37.3 ; python_version >= "3.9" and python_version < "4.0"
requests-toolbelt==0.10.1 ; python_version >= "3.9" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.9" and python_version < "4"
rfc3986==2.0.0 ; python_version >= "3.9" and python_version < "4.0"
rsa==4.9 ; python_version >= "3.9" and python_version < "4"
scipy==1.9.3 ; python_version >= "3.9" and python_version < "4.0"
seaborn==0.11.2 ; python_version >= "3.9" and python_version < "4.0"
secretstorage==3.3.3 ; python_version >= "3.9" and python_version < "4.0" and sys_platform == "linux"
sentry-sdk==1.18.0 ; python_version >= "3.9" and python_version < "4.0"
setuptools-scm==7.1.0 ; python_version >= "3.9" and python_version < "4.0"
setuptools==67.6.1 ; python_version >= "3.9" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.0 ; python_version >= "3.9" and python_version < "4.0"
sqlalchemy==1.4.47 ; python_version >= "3.9" and python_version < "4.0"
starlette-validation-uploadfile==0.1.1 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.19.1 ; python_version >= "3.9" and python_version < "4.0"
tomli==2.0.1 ; python_version >= "3.9" and python_version < "3.11"
tqdm==4.65.0 ; python_version >= "3.9" and python_version < "4.0"
twine==3.8.0 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.5.0 ; python_version >= "3.9" and python_version < "4.0"
urllib3==1.26.15 ; python_version >= "3.9" and python_version < "4"
uvicorn==0.17.6 ; python_version >= "3.9" and python_version < "4.0"
webencodings==0.5.1 ; python_version >= "3.9" and python_version < "4.0"
xlsxwriter==3.0.9 ; python_version >= "3.9" and python_version < "4.0"
zipp==3.15.0 ; python_version >= "3.9" and python_version < "4.0"

fastapi~=0.79.1
SQLAlchemy~=1.4.47
tqdm~=4.67.1
passlib~=1.7.4
pydantic~=1.10.7
numpy~=2.0.2
Levenshtein~=0.20.9
pandas~=2.3.1
python-dateutil~=2.9.0.post0
PyYAML~=6.0
orjson~=3.10.18
starlette~=0.19.1
python-dotenv~=1.1.1
uvicorn~=0.17.6
loguru~=0.7.3
swifter~=1.4.0
matplotlib~=3.10.3